Remove reusable logic
