+++ packages/react-scripts-plugin-typescript/src/index.js
@@ -1,40 +1,13 @@
 'use strict';
 
-const merge = require('lodash.merge');
-const invariant = require('invariant');
-
-// arr: [[afterExt, strExt1, strExt2, ...], ...]
-function pushExtensions(config, arr) {
-  const { resolve: { extensions } } = config;
-
-  for (const [after, ...exts] of arr) {
-    // Find the extension we want to add after
-    const index = extensions.findIndex(s => s === after);
-    invariant(
-      index !== -1,
-      `Unable to find extension ${after} in configuration.`
-    );
-    // Push the extensions into array in the order we specify
-    extensions.splice(index + 1, 0, ...exts);
-  }
-}
+const {
+  pushExtensions,
+  pushExclusiveLoader,
+} = require('react-dev-utils/plugins');
 
 function apply(config, { paths }) {
-  // Deep copy configuration
-  config = merge({}, config);
-
   pushExtensions(config, [['.js', '.tsx', '.ts']]);
-
-  const { module: { rules: [, { oneOf: rules }] } } = config;
-
-  // Find babel loader
-  const jsTransformIndex = rules.findIndex(
-    rule => rule.test.toString() === '/\\.(js|jsx)$/'
-  );
-  invariant(jsTransformIndex !== -1, 'Unable to find babel transform.');
-  // Push typescript loader after babel-loader since they're related (this
-  //  matters for ejecting)
-  rules.splice(jsTransformIndex + 1, 0, {
+  pushExclusiveLoader(config, '/\\.(js|jsx)$/', {
     test: /\.(ts|tsx)$/,
     include: paths.appSrc,
     loader: require.resolve('awesome-typescript-loader'),
@@ -43,7 +16,6 @@ function apply(config, { paths }) {
       configFileName: require.resolve('tsconfig-react-app'),
     },
   });
-
   return config;
 }
